Generalize script
Write script to cycle through all categories.  I realized that the code for this script and future scripts is going to be really awkward.  The number of unique items in any semantic fluency item is going to differ per trial and preallocating with NaNs leads to really squirrely outputs.  So probably means a lot of copy pasting and redundancy, but I definitely feel more comfortable working with this format of data.