Clean Up
Last thing before taking a break from the data.  This isn't specific to any comparison, but might make things easier in the future.  Basically I just added a col for (... sp's/rp's?) in trial 2 that correspond to the order that items were listed in trial 1.

E.g. Trial 1: ["A", "B", "C", "D"]
["A"= 1, "B"= 2, "C"= 3, "D"=4] --> [1, 2, 3, 4]

Trial 2: ["C", "A", "B", "D"]
[C= 3, A= 1, B= 2, D= 4] --> [3, 1, 2, 4]

Might not need it, but didn't take a long time and could be useful.